+++ packages/form-render/src/form-render-core/src/validator.js
@@ -63,7 +63,13 @@ const getRelatedPaths = (path, flatten) => {
   });
 };
 
-export const validateField = ({ path, formData, flatten, options }) => {
+export const validateField = ({
+  path,
+  formData,
+  flatten,
+  options,
+  formInstance = {},
+}) => {
   const paths = getRelatedPaths(path, flatten);
   // console.log('all relevant paths:', paths);
   const promiseArray = paths.map(path => {
@@ -73,7 +79,13 @@ export const validateField = ({ path, formData, flatten, options }) => {
       const singleData = get(formData, path);
       let schema = item.schema || {};
       const finalSchema = parseSchemaExpression(schema, formData, path);
-      return validateSingle(singleData, finalSchema, path, options); // is a promise
+      return validateSingle(
+        singleData,
+        finalSchema,
+        path,
+        options,
+        formInstance
+      ); // is a promise
     } else {
       return Promise.resolve();
     }
@@ -130,6 +142,7 @@ export const validateAll = ({
   formData,
   flatten,
   options, // {locale = 'cn', validateMessages = {}}
+  formInstance = {},
 }) => {
   const paths = dataToKeys(formData);
   const allPaths = getAllPaths(paths, flatten);
@@ -147,13 +160,20 @@ export const validateAll = ({
       const relatedPaths = getRelatedPaths(path, flatten);
       if (relatedPaths.length > 1) {
         const parentPath = relatedPaths[relatedPaths.length - 1];
-        if (flatten[parentPath].schema.hidden) {
+        const parentSchema = flatten[parentPath] || {};
+        if (get(parentSchema, 'schema.hidden', false)) {
           schema.hidden = true;
         }
       }
 
       const finalSchema = parseSchemaExpression(schema, formData, path);
-      return validateSingle(singleData, finalSchema, path, options); // is a promise
+      return validateSingle(
+        singleData,
+        finalSchema,
+        path,
+        options,
+        formInstance
+      ); // is a promise
     } else {
       return Promise.resolve();
     }
@@ -178,7 +198,17 @@ export const validateAll = ({
     });
 };
 
-const validateSingle = (data, schema = {}, path, options = {}) => {
+const validateSingle = (
+  data,
+  schema = {},
+  path,
+  options = {},
+  formInstance = {}
+) => {
+  // 自定义区块不做rules校验
+  if (schema.type === 'block') {
+    return Promise.resolve();
+  }
   if (schema.hidden) {
     return Promise.resolve();
   }
@@ -195,6 +225,7 @@ const validateSingle = (data, schema = {}, path, options = {}) => {
   } catch (error) {
     return Promise.resolve();
   }
+  formInstance?.setFieldValidating(path);
   let messageFeed = locale === 'en' ? en : cn;
   merge(messageFeed, validateMessages);
   validator.messages(messageFeed);
@@ -205,5 +236,8 @@ const validateSingle = (data, schema = {}, path, options = {}) => {
     })
     .catch(({ errors, fields }) => {
       return errors;
+    })
+    .finally(() => {
+      formInstance?.removeFieldValidating(path);
     });
 };