+++ packages/form-render/src/schema.d.ts
@@ -14,11 +14,12 @@ interface SchemaBase {
     | 'dateTime'
     | 'date'
     | 'time'
-    | 'upload';
+    | 'upload'
+    | (string & {});
   default: any;
   /** 是否必填，支持 `'{{ formData.xxx === "" }}'` 形式的表达式 */
   required: boolean | string;
-  placeholder: string;
+  placeholder: string | string[];
   bind: false | string | string[];
   dependencies: string[];
   min: number;