+++ packages/form-render/src/widgets/antd/index.js
@@ -1,5 +1,5 @@
-import { Checkbox, Input, InputNumber, Rate, Switch, TreeSelect } from 'antd';
 import React from 'react';
+import { Checkbox, Input, InputNumber, Rate, Switch, TreeSelect } from 'antd';
 import checkboxes from './checkboxes';
 import color from './color';
 import date from './date';