+++ packages/form-render/src/widgets/antd/map.js
@@ -1,14 +1,16 @@
-import { Collapse } from 'antd';
 import React, { useEffect, useState } from 'react';
-// import { useStore2 } from '../../hooks';
+import { Card, Collapse } from 'antd';
+import cn from 'classnames';
+import { useStore2 } from '../../form-render-core/src/hooks';
+
 const { Panel } = Collapse;
 
-export default function Map({ children, title, schema, ...rest }) {
-  const { theme, displayType, allCollapsed } = {}; // TODO!
+export default function Map({ children, title, schema }) {
   const [collapsed, setCollapsed] = useState(schema.collapsed || false);
-  // useEffect(() => {
-  //   setCollapsed(allCollapsed);
-  // }, [allCollapsed]);
+  const { theme: globalTheme, displayType: globalDisplayType } = useStore2();
+
+  const theme = schema.theme || globalTheme;
+  const displayType = schema.displayType || globalDisplayType;
 
   useEffect(() => {
     if (schema.hasOwnProperty('collapsed')) {
@@ -19,44 +21,49 @@ export default function Map({ children, title, schema, ...rest }) {
   if (!title) {
     return <div className="w-100">{children}</div>;
   }
-  // if (theme == '1') {
-  //   return (
-  //     <div className="w-100">
-  //       <div
-  //         style={{
-  //           fontSize: 17,
-  //           fontWeight: 500,
-  //           paddingBottom: 4,
-  //           borderBottom: '1px solid rgba( 0, 0, 0, .2 )',
-  //           marginBottom: 16,
-  //         }}
-  //       >
-  //         {title}
-  //       </div>
-  //       <div style={{ marginLeft: displayType == 'row' ? 0 : 12 }}>
-  //         {children}
-  //       </div>
-  //     </div>
-  //   );
-  // }
+  if (theme === 'tile') {
+    return (
+      <div className="w-100">
+        <div
+          style={{
+            fontSize: 17,
+            fontWeight: 500,
+            paddingBottom: 4,
+            borderBottom: '1px solid rgba( 0, 0, 0, .2 )',
+            marginBottom: 16,
+          }}
+        >
+          {title}
+          <span className="fr-desc ml2">
+            {schema?.description ? `( ${schema.description} )` : ''}
+          </span>
+        </div>
+        <div style={{ marginLeft: displayType == 'row' ? 0 : 12 }}>
+          {children}
+        </div>
+      </div>
+    );
+  }
 
   // 新增卡片视图
-  // if (theme == '2') {
-  //   const { id } = rest.schema;
-  //   return (
-  //     <div class="fr-theme-card-wrap">
-  //       <div>
-  //         {/* title 容器的 id，用来加锚点用 */}
-  //         <div id={id || title} class="fr-theme-card-title">
-  //           {title}
-  //         </div>
-  //         <div style={{ marginLeft: displayType == 'row' ? 0 : 12 }}>
-  //           {children}
-  //         </div>
-  //       </div>
-  //     </div>
-  //   );
-  // }
+  if (theme === 'card') {
+    return (
+      <Card
+        id={title}
+        title={
+          <>
+            {title}
+            <span className="fr-desc ml2">
+              {schema?.description ? `( ${schema.description} )` : ''}
+            </span>
+          </>
+        }
+        className="fr-theme-card-wrap"
+      >
+        {children}
+      </Card>
+    );
+  }
 
   const toggle = keyList => {
     if (keyList.length > 0) {
@@ -68,7 +75,12 @@ export default function Map({ children, title, schema, ...rest }) {
 
   return (
     <div className="w-100">
-      <Collapse activeKey={collapsed ? [] : ['1']} onChange={toggle}>
+      <Collapse
+        activeKey={collapsed ? [] : ['1']}
+        onChange={toggle}
+        bordered={theme !== 'collapse:pure'}
+        ghost={theme === 'collapse:ghost'}
+      >
         <Panel
           header={
             <span style={{ fontSize: 16, fontWeight: 500 }}>
@@ -81,13 +93,16 @@ export default function Map({ children, title, schema, ...rest }) {
           key="1"
           className="fr-collapse-object"
         >
-          {children}
+          <div
+            className={cn({
+              'fr-collapse-object-child-row': displayType === 'row',
+              'fr-collapse-object-child-column': displayType === 'column',
+            })}
+          >
+            {children}
+          </div>
         </Panel>
       </Collapse>
     </div>
   );
 }
-
-// export default function map({ children, title }) {
-//   return <div className="w-100">{children}</div>;
-// }